+++ src/Resources/config/service_knplabs.php
@@ -11,7 +11,9 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
+use Knp\DoctrineBehaviors\Contract\Provider\LocaleProviderInterface;
 use Sonata\TranslationBundle\Admin\Extension\Knplabs\TranslatableAdminExtension;
+use Sonata\TranslationBundle\Provider\Knplabs\LocaleProvider;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ReferenceConfigurator;
 
@@ -24,6 +26,14 @@ return static function (ContainerConfigurator $containerConfigurator): void {
             ->tag('sonata.admin.extension')
             ->args([
                 new ReferenceConfigurator('sonata_translation.checker.translatable'),
-                '%sonata_translation.default_locale%',
+                new ReferenceConfigurator('sonata_translation.admin.provider.request_locale_provider'),
+            ])
+
+        ->set('sonata_translation.admin.provider.knplabs_locale_provider', LocaleProvider::class)
+            ->decorate(LocaleProviderInterface::class)
+            ->args([
+                new ReferenceConfigurator('request_stack'),
+                new ReferenceConfigurator('sonata_translation.admin.provider.knplabs_locale_provider.inner'),
+                new ReferenceConfigurator('sonata_translation.admin.provider.request_locale_provider'),
             ]);
 };