+++ tests/Admin/Extension/Gedmo/TranslatableAdminExtensionTest.php
@@ -23,6 +23,7 @@ use Sonata\TranslationBundle\Admin\Extension\AbstractTranslatableAdminExtension;
 use Sonata\TranslationBundle\Admin\Extension\Gedmo\TranslatableAdminExtension;
 use Sonata\TranslationBundle\Checker\TranslatableChecker;
 use Sonata\TranslationBundle\Model\TranslatableInterface;
+use Sonata\TranslationBundle\Provider\LocaleProviderInterface;
 use Sonata\TranslationBundle\Tests\Fixtures\Model\ModelTranslatable;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\HttpFoundation\Request;
@@ -63,11 +64,18 @@ final class TranslatableAdminExtensionTest extends WebTestCase
             ->method('getManagerForClass')
             ->willReturn($objectManager);
 
+        $localeProvider = new class() implements LocaleProviderInterface {
+            public function get(): string
+            {
+                return 'es';
+            }
+        };
+
         $this->extension = new TranslatableAdminExtension(
             $translatableChecker,
             $this->translatableListener,
             $managerRegistry,
-            'es'
+            $localeProvider
         );
 
         $request = new Request();